requirements.txt
affine==2.3.0
appdirs==1.4.4
attrs==20.3.0
certifi==2020.12.5
cftime==1.4.1
chardet==4.0.0
click==7.1.2
click-plugins==1.1.1
cligj==0.7.1
cycler==0.10.0
deprecation==2.1.0
distlib==0.3.1
filelock==3.0.12
GDAL==3.2.1
h5netcdf==0.9.0
h5py==3.1.0
idna==2.10
kiwisolver==1.3.1
matplotlib==3.3.4
netCDF4==1.5.5.1
numpy==1.20.0
osgeo==0.0.0
packaging==20.9
pandas==1.2.2
pbr==5.5.1
Pillow==8.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
rasterio==1.2.0
requests==2.25.1
scipy==1.6.0
six==1.15.0
snuggs==1.4.7
stevedore==3.3.0
urllib3==1.26.3
virtualenv==20.4.2
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
wradlib==1.9.0
xarray==0.16.2
xmltodict==0.12.0
